Instructions
First, explore the students data to see how the study reached its conclusions and gain a better understanding of it. Do this by writing a query to see how the students are categorized and scored. Consider using aggregate function calls to check the summary statistics of each student group - international and domestic.

With the information you've gained, write a new query to see how the length of stay of an international student impacts the average diagnostic scores rounded to two decimal places. Order the results by descending order of the length of stay, and remember to use aliases!

The final output of your query with aliases will have a total nine observation rows and four columns: stay, average_phq, average_scs, and average_as